backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.7.1
